[doc] Fix some invalid/suspicious docstrings
Summary:
The upcoming ocamlformat has the ability to parse and format
docstrings. This requires that the docstrings conform to the ocamldoc
spec a bit more strongly. If a docstring does not parse, it is left
alone, but if it is morally ill-formed but parses by chance, it can be
reformatted incorrectly. This patch fixes the existing instances of
this problem.

Reviewed By: mbouaziz

Differential Revision: D12911937

fbshipit-source-id: 1c2eb590b