[clang] leave markers of variable initialization for pulse
Summary:
When initialising a variable via semi-exotic means, the frontend loses
the information that the variable was initialised. For instance, it
translates:

```
struct Foo { int i; };

...
Foo s = {42};
```

as:

```
s.i := 42
```

This can be confusing for backends that need to know that `s` actually
got initialised, eg pulse.

The solution implemented here is to insert of dummy call to
`__variable_initiazition`:

```
__variable_initialization(&s);
s.i := 42;
```

Then checkers can recognise that this builtin function does what its
name says.

Reviewed By: mbouaziz

Differential Revision: D12887122

fbshipit-source-id: 6e7214438