[pulse] record attributes inside memory cells instead of separately
Summary:
It turns out keeping attributes (such as invalidation facts) separate
from the memory is a bad idea and leads to loss of precision and false
positives, as seen in the new test (which previously generated a
report).

Allow me to illustrate on this example, which is a stylised version of
the issue in the added test: previously we'd have:

```
state1 = { x = 1; invalids={} }
state2 = { x = 2; invalids ={1} }
join(state1, state2) = { x = {1, 2}; invalids={{1, 2}} }
```

So even though none of the states said that `x` pointed to an invalid
location, the join state says it does because `1` and `2` have been
glommed together. The fact `x=1` from `state1` and the fact "1 is
invalid" from `state2` conspire together and `x` is now invalid even
though it shouldn't.

Instead, if we record attributes as part of the memory we get that `x`
is still valid after the join:

```
state1 = { x = (1, {}) }
state2 = { x = (2, {}) }
join(state1, state2) = { x = ({1, 2}, {}) }
```

Reviewed By: mbouaziz

Differential Revision: D12958130

fbshipit-source-id: 53dc81cc7