[starvation] suppress the number of reports suppressed
Summary:
Currently, if there are several reports on the same line, the most important one is reported together with a message containing how many reports were suppressed.
This is sometimes causing the bug hash we use believe that a report is introduced (eg if the number of suppressed reports changes).

Reviewed By: mbouaziz

Differential Revision: D13067306

fbshipit-source-id: 1cc0c6d3a