[pulse] Abstract Location Set
Summary: Instead of variable having the value of a single location on stack, we now allow variables to have multiple locations. Consequently, we also allow a memory location to point to a set of locations in the heap. We enforce a limit on a maximum number of locations in a set (currently 5).

Reviewed By: jvillard

Differential Revision: D13190876

fbshipit-source-id: 5cb5ba9a6