[inferbo] Substitute conditions of proof obligations strictly
Summary:
This diff substitutes the conditions of proof obligations strictly, so that the condition of "p!=Null" becomes bottom
when callee's p is Null.

In the non-strict substitution (which is used by default), if p's location is not found it returns the unknown location.
On the other hand, in the strict substitution (which is used only in the substitution of condition of proof obligation),
it returns bottom location.

Depends on D13415366, D13414636

Reviewed By: mbouaziz, jvillard

Differential Revision: D13415377

fbshipit-source-id: 5ae1e888e