[inferbo] Stronger typing to ensure absence of on-demand env in summary
Summary:
The `oenv` is an option.
This diff ensures that it is `Some` during the analysis and `None` when it is stored in a summary.
It could have been resolved with another type, e.g. `unit`, but an option was needed to avoid duplicating code that is generic up to some point.
The price to pay is a parametric type.

Reviewed By: skcho

Differential Revision: D13825418

fbshipit-source-id: 71824609d