[travis] use ocaml-ci-scripts to build infer
Summary:
With this change Travis will build infer on Linux + OSX using the opam file, in
Java mode only to save us from having to compile clang. This doesn't try to run
the tests yet because 1) building infer takes almost all the alloted time so
OSX times out if we ask it to do more; 2) there's some error in Travis when the
tests are run so more work is needed.

I had to change the opam file a bit to reflect some dependency constraints
since the opam build ignores the opam.locked file (which is good because it
will allow us to detect when such incompatibilities arise).

Reviewed By: katiejots

Differential Revision: D13971408

fbshipit-source-id: d51caa66f