[inferbo] Avoid precision-losing pruning
Summary:
In this diff, it avoids a precision-losing pruning, which was needed
to keep effects of assume commands.

```
unsigned int c = a + b; // (1)
if (c > 0) { // (2)
  char result[c];
  result[c - 1] = 0; // (4)
}
```

For example, in the example, `c` is assigned by `[a+b,a+b]` at (1),
then it tried to prune the lower bound of `c` to 1 at (2) while losing
precision, in order to say `c - 1` at (4) is safe in terms of integer
underflow.  Instead, it could not say that `c - 1` is smaller than `c`
in the buffer access, because the former is analyzed to `[0,a+b-1]` and
the latter `[1,a+b]` at (4).

Now, the situation has changed.  By adopting conditional proof
obligation (D13749914), the FP of integer overflow can be suppressed
without the precision-losing pruning.

Reviewed By: mbouaziz

Differential Revision: D14122770

fbshipit-source-id: 634744e99