[pulse] perf win: separate edges from attrs in memory
Summary:
Before: the abstract state represents heap addresses as a single map
from addresses to edges + attributes.

After: the heap is made of 2 maps: one mapping addresses to edges, and
one mapping an address to its attributes.

It turns out that edges and attributes are often not updated at the same
time, so keeping them in the same map was causing pressure on the OCaml
gc.

Reviewed By: mbouaziz

Differential Revision: D14147991

fbshipit-source-id: 6713eeb3c