[inferbo] Revise substitution of array block
Summary:
Given a pointer-typed parameter, Inferbo assumes that it is an array
block.  However, when a pointer is given as an actual parameter, it
failed the substitution of the array block value of the parameter, thus
which made some return values to bottom unexpectedly.

This diff revises the substitution of array block, so it can
substitute array block values with actual pointers correctly when it
is possible.

Reviewed By: mbouaziz

Differential Revision: D14663475

fbshipit-source-id: 0477de1ba