[pulse] improve error messages and traces
Summary:
The previous message formatting had regressed and produced non-sensical messages.

More importantly, remove template parameters from error messages to
trigger the heuristic in `InferPrint` that deduplicates errors that are
on the same line with the same error type and message. Without this we
get hundreds of reports that correspond to as many instantiations of the
same code.

Reviewed By: ngorogiannis

Differential Revision: D14747979

fbshipit-source-id: 3c4aad2b1