[pulse] do not create `&` back-edges eagerly
Summary:
This mostly doesn't make sense. The only thing this would have been good
for was to give the most accurate result on access paths such as
`*(&(x.f))`, but these are normalised anyway (into `x.f`) so we actually
never see these. That said there might be some use to some similar logic
in the future, but in the meantime let's delete the current feature as
it wasn't thought through.

Reviewed By: ezgicicek

Differential Revision: D14753492

fbshipit-source-id: 597cec027