[pulse] be more careful about what to consider as a variable going out of scope
Summary:
The heuristic to detect variables going out of scope was to detect any
access expression passed as argument to an injected destructor call.
However destructor calls are also injected in destructor bodies to
destruct each field of an object, so the heuristic would detect fields
going out of scope, which, erm, doesn't make sense. Limit the heuristic
to local program variables.

Reviewed By: jberdine

Differential Revision: D14771454

fbshipit-source-id: ffa3c9fe3