[report] restore missing "could be null and is dereferenced" message for nullable dereference
Summary:
We get messages like " object returned by `getArguments()` at line 101."
instead of " object returned by `getArguments()` could be null and is
dereferenced at line 101.". Tracking it down, it happens for
nullable-looking values, but I don't know why.

It seems that something regressed but I couldn't track it down.

So, just generate the error message in the same way as for non-nullable
objects in this case to fix the non-sensical message.

Reviewed By: jeremydubreil

Differential Revision: D14972325

fbshipit-source-id: 2a97501cc