[pulse] traces record how values were constructed
Summary:
Before: the trace would explain how a value was invalidated and
accessed, but not how the value that was invalidated had been
constructed.

Now: `PulseTrace.t` records breadcrumbs of how the value was constructed
in addition to the interproc "action" trace leading to the invalidation
or access action.

Concretely:

```
void bad(X &x) {
  X *y = x;
  X *z = x;
  delete y;
  access(z);
}
```

will produce the trace:

Invalidation part:
y = x
delete y

Access part:
z = x
access(z)
access to z->f inside of access(z)

Before this diff the "Access part" would be missing the "z = x" part of
the trace, so it might be confusing why `z` has anything to do with `y`.

However, such "breadcrumbs" are not recorded in the inter-procedural
part, only the sequence of calls is. This is a trade-off for simplicity,
maybe it's enough for developers maybe it isn't, we'll find out later.

Reviewed By: jberdine

Differential Revision: D15354438

fbshipit-source-id: 8d0aed717