[sledge] Avoid calling Llvm.string_of_llvalue on instructions
Summary:
Llvm.string_of_llvalue, which just calls llvm::Value::print, is
extremely slow when called on instructions or functions. In these
cases, it initializes metadata slots for everything in the *parent
module* of the instruction or function being printed, on every
call. This is ridiculously slow, don't do it.

Reviewed By: kren1

Differential Revision: D15511376

fbshipit-source-id: 658eeccab