[buckjava2] Move genrule capture integration logic from shell to OCaml
Summary:
Move genrule capture integration logic from shell to OCaml.

Also, stop relying on side-effects of buck compilation for constructing the infer-deps.txt file used for merging.  Now this is obtained by passing `--show-output` to buck, which spits out the `buck-out` output paths to the targets we asked to build.

Reviewed By: ezgicicek

Differential Revision: D15715608

fbshipit-source-id: 8fa896ba6