[inefficient-keyset-iterator] New checker for finding inefficient keySet iterator
Summary:
This is a simple checker that identifies inefficient uses of `keySet` iterator where (not only the key but also) the value is accessed via `get(key)`. It is more efficient to use `entrySet` iterator which already returns both key-value pairs. This optimization would get rid of many extra lookups which can be expensive.

We simply traverse the CFG starting from the loop head upwards and pick up the map that is iterated over. Then, we check in the loop nodes if there is a call to `get(...)` over this map. If, so we report.

Reviewed By: ngorogiannis

Differential Revision: D15737779

fbshipit-source-id: 702465b4e